Reworked initial structure
