--- .eslintignore
@@ -1,10 +1,0 @@
-# /node_modules/* in the project root is ignored by default
-# build artefacts
-dist/*
-coverage/*
-# data definition files
-**/*.d.ts
-# 3rd party libs
-/src/public/
-# custom definition files
-/src/types/
