+++ package.json
@@ -1,27 +1,24 @@
 {
-  "name": "cs-4500-chemistry-nightmare",
+  "name": "chemistry-nightmare",
   "version": "0.1.0",
   "description": "A project for a class",
+  "main": "app.js",
   "scripts": {
-    "build": "npm run build-clean && npm run build-sass && npm run build-ts && npm run lint && npm run copy-static-assets",
-    "build-clean": "rimraf dist",
-    "build-sass": "node-sass src/public/css/main.scss dist/public/css/main.css",
-    "build-ts": "tsc",
-    "copy-static-assets": "ts-node copyStaticAssets.ts",
-    "lint": "tsc --noEmit && eslint \"**/*.{js,ts}\" --quiet --fix",
-    "serve": "node dist/server.js",
-    "serve-debug": "nodemon --inspect dist/server.js",
-    "start": "npm run serve"
+    "start": "npm run start:prod",
+    "start:dev": "nodemon app.js",
+    "start:prod": "node app.js"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/EmWhite-a11y/CS4500ChemistryNightmare.git"
   },
   "keywords": [],
-  "contributors": [{
-    "name": "Jared Diehl",
-    "email": "jmddnb@umsystem.edu"
-  }],
+  "contributors": [
+    {
+      "name": "Jared Diehl",
+      "email": "jmddnb@umsystem.edu"
+    }
+  ],
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/EmWhite-a11y/CS4500ChemistryNightmare/issues"
@@ -29,34 +26,13 @@
   "homepage": "https://github.com/EmWhite-a11y/CS4500ChemistryNightmare#readme",
   "dependencies": {
     "body-parser": "^1.19.0",
-    "dotenv": "^8.2.0",
-    "errorhandler": "^1.5.1",
+    "compression": "^1.7.4",
     "express": "^4.17.1",
     "pug": "^3.0.0",
-    "serve-favicon": "^2.5.0",
-    "socket.io": "^2.3.0",
-    "winston": "^3.3.3"
+    "socket.io": "^2.3.0"
   },
   "devDependencies": {
-    "@types/body-parser": "^1.19.0",
-    "@types/errorhandler": "^1.5.0",
-    "@types/eslint": "^7.2.4",
-    "@types/express": "^4.17.8",
-    "@types/jquery": "^3.5.4",
-    "@types/node": "^14.14.6",
-    "@types/node-sass": "^4.11.1",
-    "@types/pug": "^2.0.4",
-    "@types/serve-favicon": "^2.5.1",
-    "@types/shelljs": "^0.8.8",
-    "@types/winston": "^2.4.4",
-    "@typescript-eslint/eslint-plugin": "^4.6.1",
-    "@typescript-eslint/parser": "^4.6.1",
     "eslint": "^7.12.1",
-    "node-sass": "^5.0.0",
-    "nodemon": "^2.0.6",
-    "rimraf": "^3.0.2",
-    "shelljs": "^0.8.4",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.5"
+    "nodemon": "^2.0.6"
   }
 }
