+++ public/js/game.js
@@ -1,0 +1,1 @@
+var socket = io();
