+++ public/js/main.js
@@ -1,0 +1,3 @@
+$(function() {
+    console.log('Initialized')
+})
