--- src/controllers/home.ts
@@ -1,11 +1,0 @@
-import { Request, Response } from "express";
-
-/**
- * Home page.
- * @route GET /
- */
-export const index = (req: Request, res: Response) => {
-    res.render("home", {
-        title: "Home"
-    });
-};
