--- src/public/js/main.ts
@@ -1,3 +1,0 @@
-$(function() {
-    console.log("Scripts initialized");
-});
