--- src/server.ts
@@ -1,10 +1,0 @@
-import app from "./app";
-import errorHandler from "errorhandler";
-
-app.use(errorHandler());
-
-const server = app.listen(app.get("port"), () => {
-    console.log("App is running at http://localhost:%d in %s mode", app.get("port"), app.get("env"));
-});
-
-export default server;
