--- src/util/secrets.ts
@@ -1,10 +1,0 @@
-import dotenv from "dotenv";
-import fs from "fs";
-import logger from "./logger";
-
-if (fs.existsSync(".env")) {
-    logger.debug("Using .env file to supply config environment variables");
-    dotenv.config({ path: ".env" });
-}
-
-export const ENVIRONMENT = process.env.NODE_ENV;
