+++ views/game.pug
@@ -1,0 +1,12 @@
+extends layout
+
+block styles
+    link(rel="stylesheet", href="/css/game.css")
+
+block content
+    | Game #{id}
+
+block scripts
+    script(src="/socket.io/socket.io.js")
+    script(src="https://cdnjs.cloudflare.com/ajax/libs/phaser/3.24.1/phaser.min.js", integrity="sha512-HY/qPTJqpvoNGRufGPn4Tl7duHSam8NfrfTZUHyn5M39oEBcq8p6IeGBE9e+AupikS4Utt+uJhSgG3FYZp/yxA==", crossorigin="anonymous")
+    script(src="/js/game.js")
