+++ views/home.pug
@@ -1,4 +1,10 @@
 extends layout
 
+block styles
+    link(rel="stylesheet", href="/css/home.css")
+
 block content
     h1 Chemistry Nightmare
+
+block scripts
+    script(src="/js/home.js")
