+++ views/layout.pug
@@ -9,12 +9,10 @@ html(lang='')
         meta(name='theme-color', content='#33b864')
         link(rel='stylesheet', href='https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css', integrity='sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z', crossorigin='anonymous')
         link(rel='stylesheet', href='/css/main.css')
+        block styles
     
     body
-        include partials/header
-
         .container
-            include partials/flash
             block content
         
         include partials/footer
@@ -22,3 +20,4 @@ html(lang='')
         script(src='https://code.jquery.com/jquery-3.5.1.min.js', integrity='sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=', crossorigin='anonymous')
         script(src='https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.bundle.min.js', integrity='sha384-LtrjvnR4Twt/qOuYxE721u19sVFLVSA4hf/rRt6PrZTmiPltdZcI7q7PXQBYTKyf', crossorigin='anonymous')
         script(src='/js/main.js')
+        block scripts
