--- views/partials/header.pug
@@ -1,2 +1,0 @@
-header
-    | Header
